+++ src/main/java/com/raoulvdberge/refinedstorage/apiimpl/network/node/NetworkNode.java
@@ -6,6 +6,8 @@ import com.raoulvdberge.refinedstorage.api.network.INetworkNodeVisitor;
 import com.raoulvdberge.refinedstorage.api.network.node.INetworkNode;
 import com.raoulvdberge.refinedstorage.api.util.IWrenchable;
 import com.raoulvdberge.refinedstorage.apiimpl.API;
+import com.raoulvdberge.refinedstorage.integration.mcmp.IntegrationMCMP;
+import com.raoulvdberge.refinedstorage.integration.mcmp.RSMCMPAddon;
 import com.raoulvdberge.refinedstorage.tile.TileBase;
 import com.raoulvdberge.refinedstorage.tile.config.RedstoneMode;
 import net.minecraft.block.state.IBlockState;
@@ -191,8 +193,9 @@ public abstract class NetworkNode implements INetworkNode, INetworkNodeVisitor, 
         return direction;
     }
 
+    // @todo: move this data to the network node
     public void resetDirection() {
-        this.direction = ((TileBase) world.getTileEntity(pos)).getDirection();
+        this.direction = ((TileBase) (IntegrationMCMP.isLoaded() ? RSMCMPAddon.unwrapTile(world, pos) : world.getTileEntity(pos))).getDirection();
     }
 
     @Nullable