+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Refined Storage Changelog
 
 ### 1.5.8
+- Updated Forge to 2400 (raoulvdberge)
 - Updated Storage Drawers API (raoulvdberge)
 - Fixed bug where disks have to be re-inserted in the Disk Drive in order to work again after rejoining a chunk (raoulvdberge)
 - Fixed bug where items inserted in Storage Drawers through External Storage with a Drawer Controller wouldn't respect drawer priority rules (raoulvdberge)